refactor: extract UI module from loom_io into dedicated ui package
