+++ src/ui/__init__.py
@@ -1,0 +1,12 @@
+"""
+src/ui
+UI components for terminal interaction (progress, input, console helpers).
+
+Exports:
+- UI: main interface for progress-safe user I/O
+"""
+
+from .ui import UI
+
+__all__ = ["UI"]
+
